+++ src/main/java/com/digital_tok/dto/request/AuthRequestDTO.java
@@ -1,0 +1,32 @@
+package com.digital_tok.dto.request;
+
+import io.swagger.v3.oas.annotations.media.Schema;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+
+public class AuthRequestDTO {
+    // 1. 회원가입 요청
+    @Getter
+    public static class JoinDto {
+        private String email;       // 필수
+        private String password;    // 필수
+        private String nickname;    // 필수
+        private String name;        // 선택
+        private String phoneNumber; // 선택
+    }
+
+    // 2. 로그인 요청
+    @Getter
+    public static class LoginDto {
+        private String email;       // 필수
+        private String password;    // 필수
+    }
+
+    // 3. 로그아웃 요청 (Body에 있는 Refresh Token 값)
+    @Getter
+    public static class LogoutDto {
+        private String refreshToken; // 필수 (DB/Redis 삭제용)
+    }
+}
